Got rid of signaler_actions.go. Moved GetHandler into repo_interface.go
